# docs/swagger.yaml
basePath: /
definitions:
  domain.CreateListingRequest:
    properties:
      city:
        type: string
      description:
        type: string
      images:
        items:
          type: string
        type: array
      latitude:
        type: number
      longitude:
        type: number
      preview_url:
        type: string
      status:
        enum:
        - search
        - give
        type: string
      street:
        type: string
      title:
        type: string
    required:
    - city
    - latitude
    - longitude
    - status
    - title
    type: object
...
  domain.Draft:
    properties:
      city:
        type: string
      description:
        type: string
      id:
        type: string
      images:
        items:
          type: string
        type: array
      latitude:
        type: number
      longitude:
        type: number
      owner_id:
        type: string
      preview_url:
        type: string
      status:
        type: string
      street:
        type: string
      title:
        type: string
    type: object
  domain.FileRequest:
    properties:
      content_type:
        type: string
      file_name:
        minLength: 1
        type: string
    required:
    - content_type
    - file_name
    type: object
  domain.FileUploadResponse:
    properties:
      file_name:
        type: string
      presigned_url:
        type: string
      public_url:
        type: string
    type: object
  domain.Listing:
    properties:
      city:
        type: string
      created_at:
        type: string
      description:
        type: string
      distance_m:
        type: number
      id:
        type: string
      images:
        items:
          type: string
        type: array
      latitude:
        type: number
      longitude:
        type: number
      owner:
        $ref: '#/definitions/domain.Owner'
      preview_url:
        type: string
      status:
        type: string
      street:
        type: string
      title:
        type: string
    type: object
  domain.Owner:
    properties:
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      rating:
        type: number
      surname:
        type: string
    type: object
...
  domain.UserRegisterRequest:
    properties:
      email:
        type: string
      name:
        minLength: 3
        type: string
      password:
        minLength: 6
        type: string
      phone:
        minLength: 6
        type: string
      surname:
        minLength: 3
        type: string
    required:
    - email
    - name
    - password
    - phone
    - surname
    type: object
info:
  contact: {}
  description: API для виртуальной доски объявлений между студентами
  title: StudShare API
  version: "1.0"
paths:
  /api/files/upload:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные о файле
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.FileRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.FileUploadResponse'
        "400":
          description: Некорректный запрос
          schema:
            type: string
        "401":
          description: Неавторизован
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Генерация URL для загрузки файла
      tags:
      - Files
  /api/listings:
    delete:
      parameters:
      - description: ID объявления
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            type: string
        "403":
          description: Нет доступа
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Удалить объявление
      tags:
      - Listings
    get:
      parameters:
      - description: Статус объявления
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Listing'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить все объявления (по статусу)
      tags:
      - Listings
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные объявления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.CreateListingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Создано
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            type: string
        "401":
          description: Неавторизован
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Создание объявления
      tags:
      - Listings
    put:
      consumes:
      - application/json
      parameters:
      - description: ID объявления
        in: query
        name: id
        required: true
        type: string
      - description: Объявление
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.Listing'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            type: string
        "401":
          description: Неавторизован
          schema:
            type: string
        "403":
          description: Нет доступа
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Обновить объявление
      tags:
      - Listings
  /api/listings/:
    get:
      parameters:
      - description: ID объявления
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Listing'
        "404":
          description: Объявление не найдено
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить объявление по ID
      tags:
      - Listings
  /api/listings/near:
    get:
      parameters:
      - description: Широта
        in: query
        name: lat
        required: true
        type: number
      - description: Долгота
        in: query
        name: lon
        required: true
        type: number
      - description: Радиус поиска (км)
        in: query
        name: radius
        required: true
        type: number
      - description: Статус
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Listing'
            type: array
        "400":
          description: Некорректные координаты
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить объявления поблизости
      tags:
      - Listings
  /auth/logout:
    post:
      description: Инвалидирует JWT токен
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Выход из аккаунта
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя
      parameters:
      - description: Данные пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: created
          schema:
            type: string
        "400":
          description: invalid request
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Регистрация пользователя
      tags:
      - auth
  /drafts/:
    delete:
      parameters:
      - description: ID черновика
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удалить черновик по ID
      tags:
      - drafts
    get:
      parameters:
      - description: ID черновика
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Draft'
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить черновик по ID
      tags:
      - drafts
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные черновика
        in: body
        name: draft
        required: true
        schema:
          $ref: '#/definitions/domain.Draft'
      produces:
      - application/json
      responses:
        "201":
          description: created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Создать черновик объявления
      tags:
      - drafts
  /drafts/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Draft'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить все черновики пользователя
      tags:
      - drafts
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"